made a few improvements based on solution repo recommendations
