+++ starter.rb
@@ -1,0 +1,4 @@
+
+require "./dictionary_ui.rb"
+
+start = DictionaryUI.new
